Feature/s3 simplify onboarding (#2299)
* Updated S3 Log Source onboard fields (#2183)

* export goToStep from wizard context

* Added condition for skipping step in S3SourceConfigurationPanel

* Transform prefix & logTypes to S3PrefixLogTypes

* Changed order of KMS field

* Fixes based on feedback

* Quick wording fix

* Fix creating s3 log source tests

* Fixes tests for EditLogSource

* Added additonal logType for combobox

* Added type to validation

* Renamed var for skipping cfn step

* Optimizing changes

* Fix breaking tes

* Added a test case for removing an existing prefix

* Added React.memo for flatten logTypes

* Fixed issue where button was active before query execution

* Updated snapshots

* Fixed dependicies for memo

* Feedback fixes

* Updated LinkButton

* Updated broken tests

* Removed s3PrefixLogTypes from GetCfnTemplate

* Backend: Allow users to specify multiple s3 prefix->logtypes mappings per s3 log source (#2232)

* Allow multiple s3 prefixes per s3 source

* Cleanup prefix parameter from log-onboarding CFN template

Now that users can define multiple s3 prefixes per onboarded
bucket (log source), this commit removes the prefix filter from the
`s3:GetObject` permission for the log processing role so that users
can eventually update prefixes in an s3 source without having to
update the CFN template.

"Eventually" is the key word here though. Due to backwards compatibility,
users that already have deployed a stack that includes a prefix restriction,
need to update their stack to remove it. For now, we have to prompt all users
to update their stack.

In order to allow some users to not update their CFN stack, we have to change
the current FE<>BE flow so that the backend can check whether the installed
stack needs update (by querying AWS) and return a flag to the frontend accordingly.

* Add unit test for backwards compatibility

* Fix json fields to match graphql schema

* Get unique logtypes for s3 source

* Add validation for unique prefixes per source and prefix/bucket pairs

* Narrow down prefixes in the self registered sources for Cloudtrail, GuardDuty, ALB, s3 access logs

* For s3 sources in the log processor, use the log types of the longest prefix that matches the input object key

Co-authored-by: Alex Mylonas <alex.a.mylonas@gmail.com>